Updated review form - totals rudimentally calculated and available via the Hotel Info Service.
The form is complete but I still have some work to do to make the total charge amount readily available.  I need to move the final charge amount to the reservation service.

Also some refactoring on Hotel-Info.ts to clean up the service as much as possible with the least amount of potential breakage (I utilized Hotel.ts). Ultimately Hotel-Info.ts and Hotel.ts should be merged to be all-inclusive, but may not have time for it.